+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add module-level `load`, `loads`, `dump`, and `dumps` functions.
+
 ## 1.1.4 (19 April 2026)
 
 - New entries appended/inserted into an array of tables now copy the indent